ai/requirements.txt
openai==1.34.0
fastapi==0.111.0
pydantic-extra-types==2.8.0